Further refactoring and fixed loading screen bug
