+++ Assets/Scripts/HistorialReplayMission.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 321ac4011c547f24bb76b4eb001a6a9e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
